requirements.txt
gpx-converter
fitparse
pytz
pandas
matplotlib
pillow
